chore: split PTY restart-decision tests to respect the 500-line cap
PR #570 grew model_input_lifecycle_pty_lifecycle_test.go to 606 lines,
past the hard 500-line lint gate (the executor's authorized --no-verify
for an unrelated environmental artifact masked it, and review missed
it). Move the restart-decision tests and their livePTYTab helper into
model_input_lifecycle_pty_restart_decision_test.go (128 lines); the
original returns to 487. No test content changes; center suite green.
